Redesign Project plugin. Deleting Project code from backend
