+++ meson.build
@@ -1,5 +1,5 @@
 project('MiAZ',
-         version : '0.1.22+build.9',
+         version : '0.1.22+build.21',
     meson_version: '>= 1.5.1',
   default_options: [ 'warning_level=2',
                    ],
